Extract tests related to the `Load more` logic into a separated test class
